mkerlimage: increase image size from 1GB to 4GB
With FreeBSD 13.0 the root filesystem this script produces is about
2.5GB large, so the old size of 1GB no longer works. That said, FreeBSD
13.0 on ERLite-3 soon after booting suffers a kernel trap trying to
handle an unaligned memory access, but to observe that one needs to have
a bootable image in the first place.

My ERLite-3 device does not have the original USB storage, so I cannot
confirm its exact size. Use 4GB as that's the size mentioned on the
accompanying web-page.